Fix bugs debit and credit notes
