+++ erpnext/accounts/doctype/credit_note_cxc/credit_note_cxc.json
@@ -1,31 +1,35 @@
 {
  "autoname": "naming_series:",
- "creation": "2021-04-23 16:32:41.579580",
+ "creation": "2021-04-27 13:18:25.098929",
  "doctype": "DocType",
  "editable_grid": 1,
  "engine": "InnoDB",
  "field_order": [
   "naming_series",
-  "status",
-  "supplier",
-  "transaction_number",
-  "cai",
+  "party_type",
+  "customer",
+  "rtn",
   "column_break_1",
   "posting_date",
   "due_date",
   "reason_credit_note",
+  "cai_info_section",
+  "cai",
+  "due_date_cai",
+  "authorized_range",
   "section_break_1",
+  "references",
+  "total_references",
+  "section_break_2",
   "amount",
   "taxes",
   "section_break_3",
-  "company",
-  "affects_purchases_book",
-  "affects_purchases_month",
+  "affects_sale_book",
+  "affects_sales_month",
   "column_break_2",
-  "isv_18",
   "isv_15",
-  "document_balance",
-  "outstanding_amount",
+  "isv_18",
+  "amount_total",
   "amended_from"
  ],
  "fields": [
@@ -35,22 +39,26 @@
    "fieldtype": "Select",
    "label": "Series",
    "no_copy": 1,
-   "options": "CN-CXC-.YYYY.-",
+   "options": "CN-CXP-.YYYY.-",
    "print_hide": 1,
    "reqd": 1,
    "set_only_once": 1
   },
   {
-   "default": "Draft",
-   "fieldname": "status",
-   "fieldtype": "Select",
+   "fieldname": "customer",
+   "fieldtype": "Link",
    "in_list_view": 1,
-   "label": "Status",
-   "no_copy": 1,
-   "options": "\nDraft\nSubmitted\nPaid\nUnpaid\nOverdue\nCancelled",
-   "print_hide": 1,
-   "read_only": 1,
+   "label": "Customer",
+   "options": "Customer",
    "reqd": 1
+  },
+  {
+   "fetch_from": "customer.rtn",
+   "fetch_if_empty": 1,
+   "fieldname": "rtn",
+   "fieldtype": "Data",
+   "label": "RTN",
+   "read_only": 1
   },
   {
    "fieldname": "column_break_1",
@@ -69,14 +77,43 @@
    "reqd": 1
   },
   {
-   "fieldname": "cai",
-   "fieldtype": "Data",
-   "label": "Cai",
+   "default": "Customer",
+   "fieldname": "party_type",
+   "fieldtype": "Link",
+   "label": "Party Type",
+   "options": "DocType",
+   "read_only": 1,
+   "set_only_once": 1
+  },
+  {
+   "fieldname": "reason_credit_note",
+   "fieldtype": "Link",
+   "label": "Reason Credit Note",
+   "options": "Reason Debit And Credit Note",
    "reqd": 1
   },
   {
    "fieldname": "section_break_1",
-   "fieldtype": "Section Break"
+   "fieldtype": "Section Break",
+   "label": "Document to which it applies"
+  },
+  {
+   "fieldname": "references",
+   "fieldtype": "Table",
+   "label": "References",
+   "options": "Note References"
+  },
+  {
+   "fieldname": "total_references",
+   "fieldtype": "Currency",
+   "label": "Total References",
+   "options": "Currency",
+   "read_only": 1
+  },
+  {
+   "fieldname": "section_break_2",
+   "fieldtype": "Section Break",
+   "label": "Totals and tax"
   },
   {
    "fieldname": "amount",
@@ -91,61 +128,16 @@
    "fieldtype": "Column Break"
   },
   {
-   "fieldname": "document_balance",
-   "fieldtype": "Currency",
-   "label": "Document Balance",
-   "options": "Currency",
-   "read_only": 1
-  },
-  {
-   "fieldname": "outstanding_amount",
-   "fieldtype": "Currency",
-   "label": "Outstanding Amount",
-   "options": "Currency",
-   "read_only": 1
-  },
-  {
-   "fieldname": "reason_credit_note",
-   "fieldtype": "Link",
-   "label": "Reason Credit Note",
-   "options": "Reason Debit And Credit Note"
-  },
-  {
-   "fieldname": "transaction_number",
-   "fieldtype": "Data",
-   "label": "Transaction Number",
-   "reqd": 1
-  },
-  {
-   "fieldname": "company",
-   "fieldtype": "Link",
-   "in_standard_filter": 1,
-   "label": "Company",
-   "options": "Company",
-   "print_hide": 1,
-   "read_only": 1,
-   "remember_last_selected_value": 1,
-   "reqd": 1
+   "default": "0",
+   "fieldname": "affects_sale_book",
+   "fieldtype": "Check",
+   "label": "Affects Sales Book"
   },
   {
    "default": "0",
-   "fieldname": "affects_purchases_book",
+   "fieldname": "affects_sales_month",
    "fieldtype": "Check",
-   "label": "Affects Purchases Book"
-  },
-  {
-   "default": "0",
-   "fieldname": "affects_purchases_month",
-   "fieldtype": "Check",
-   "label": "Affects Purchases Month"
-  },
-  {
-   "fieldname": "supplier",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Supplier",
-   "options": "Supplier",
-   "reqd": 1
+   "label": "Affects Sales Month"
   },
   {
    "fieldname": "amended_from",
@@ -157,14 +149,40 @@
    "read_only": 1
   },
   {
+   "collapsible": 1,
+   "fieldname": "cai_info_section",
+   "fieldtype": "Section Break",
+   "label": "Cai Info"
+  },
+  {
+   "fieldname": "cai",
+   "fieldtype": "Data",
+   "label": "Cai",
+   "read_only": 1
+  },
+  {
+   "fieldname": "due_date_cai",
+   "fieldtype": "Date",
+   "label": "Due Date Cai",
+   "read_only": 1
+  },
+  {
+   "fieldname": "authorized_range",
+   "fieldtype": "Data",
+   "label": "Authorized Range",
+   "read_only": 1
+  },
+  {
    "fieldname": "taxes",
    "fieldtype": "Table",
    "label": "Taxes",
-   "options": "Multiple Taxes",
-   "reqd": 1
+   "options": "Multiple Taxes"
   },
   {
-   "fetch_from": "taxes.base_isv * (taxes.tax_rate/100)",
+   "fieldname": "section_break_3",
+   "fieldtype": "Section Break"
+  },
+  {
    "fieldname": "isv_15",
    "fieldtype": "Currency",
    "label": "Isv 15%",
@@ -172,7 +190,6 @@
    "read_only": 1
   },
   {
-   "fetch_from": "taxes.base_isv * (taxes.tax_rate/100)",
    "fieldname": "isv_18",
    "fieldtype": "Currency",
    "label": "Isv 18%",
@@ -180,12 +197,15 @@
    "read_only": 1
   },
   {
-   "fieldname": "section_break_3",
-   "fieldtype": "Section Break"
+   "fieldname": "amount_total",
+   "fieldtype": "Currency",
+   "label": "Amount Total",
+   "options": "Currency",
+   "read_only": 1
   }
  ],
  "is_submittable": 1,
- "modified": "2021-05-04 10:30:23.097837",
+ "modified": "2021-05-07 10:40:55.288941",
  "modified_by": "Administrator",
  "module": "Accounts",
  "name": "Credit Note CXC",
