+++ erpnext/accounts/doctype/credit_note_cxp/credit_note_cxp.json
@@ -1,35 +1,31 @@
 {
  "autoname": "naming_series:",
- "creation": "2021-04-27 13:18:25.098929",
+ "creation": "2021-04-23 16:32:41.579580",
  "doctype": "DocType",
  "editable_grid": 1,
  "engine": "InnoDB",
  "field_order": [
   "naming_series",
-  "party_type",
-  "customer",
-  "rtn",
+  "status",
+  "supplier",
+  "transaction_number",
+  "cai",
   "column_break_1",
   "posting_date",
   "due_date",
   "reason_credit_note",
-  "cai_info_section",
-  "cai",
-  "due_date_cai",
-  "authorized_range",
   "section_break_1",
-  "references",
-  "total_references",
-  "section_break_2",
   "amount",
   "taxes",
   "section_break_3",
-  "affects_sale_book",
-  "affects_sales_month",
+  "company",
+  "affects_purchases_book",
+  "affects_purchases_month",
   "column_break_2",
-  "isv_15",
   "isv_18",
-  "amount_total",
+  "isv_15",
+  "document_balance",
+  "outstanding_amount",
   "amended_from"
  ],
  "fields": [
@@ -39,26 +35,22 @@
    "fieldtype": "Select",
    "label": "Series",
    "no_copy": 1,
-   "options": "CN-CXP-.YYYY.-",
+   "options": "CN-CXC-.YYYY.-",
    "print_hide": 1,
    "reqd": 1,
    "set_only_once": 1
   },
   {
-   "fieldname": "customer",
-   "fieldtype": "Link",
+   "default": "Draft",
+   "fieldname": "status",
+   "fieldtype": "Select",
    "in_list_view": 1,
-   "label": "Customer",
-   "options": "Customer",
+   "label": "Status",
+   "no_copy": 1,
+   "options": "\nDraft\nSubmitted\nPaid\nUnpaid\nOverdue\nCancelled",
+   "print_hide": 1,
+   "read_only": 1,
    "reqd": 1
-  },
-  {
-   "fetch_from": "customer.rtn",
-   "fetch_if_empty": 1,
-   "fieldname": "rtn",
-   "fieldtype": "Data",
-   "label": "RTN",
-   "read_only": 1
   },
   {
    "fieldname": "column_break_1",
@@ -77,43 +69,14 @@
    "reqd": 1
   },
   {
-   "default": "Customer",
-   "fieldname": "party_type",
-   "fieldtype": "Link",
-   "label": "Party Type",
-   "options": "DocType",
-   "read_only": 1,
-   "set_only_once": 1
-  },
-  {
-   "fieldname": "reason_credit_note",
-   "fieldtype": "Link",
-   "label": "Reason Credit Note",
-   "options": "Reason Debit And Credit Note",
+   "fieldname": "cai",
+   "fieldtype": "Data",
+   "label": "Cai",
    "reqd": 1
   },
   {
    "fieldname": "section_break_1",
-   "fieldtype": "Section Break",
-   "label": "Document to which it applies"
-  },
-  {
-   "fieldname": "references",
-   "fieldtype": "Table",
-   "label": "References",
-   "options": "Note References"
-  },
-  {
-   "fieldname": "total_references",
-   "fieldtype": "Currency",
-   "label": "Total References",
-   "options": "Currency",
-   "read_only": 1
-  },
-  {
-   "fieldname": "section_break_2",
-   "fieldtype": "Section Break",
-   "label": "Totals and tax"
+   "fieldtype": "Section Break"
   },
   {
    "fieldname": "amount",
@@ -128,16 +91,61 @@
    "fieldtype": "Column Break"
   },
   {
-   "default": "0",
-   "fieldname": "affects_sale_book",
-   "fieldtype": "Check",
-   "label": "Affects Sales Book"
+   "fieldname": "document_balance",
+   "fieldtype": "Currency",
+   "label": "Document Balance",
+   "options": "Currency",
+   "read_only": 1
+  },
+  {
+   "fieldname": "outstanding_amount",
+   "fieldtype": "Currency",
+   "label": "Outstanding Amount",
+   "options": "Currency",
+   "read_only": 1
+  },
+  {
+   "fieldname": "reason_credit_note",
+   "fieldtype": "Link",
+   "label": "Reason Credit Note",
+   "options": "Reason Debit And Credit Note"
+  },
+  {
+   "fieldname": "transaction_number",
+   "fieldtype": "Data",
+   "label": "Transaction Number",
+   "reqd": 1
+  },
+  {
+   "fieldname": "company",
+   "fieldtype": "Link",
+   "in_standard_filter": 1,
+   "label": "Company",
+   "options": "Company",
+   "print_hide": 1,
+   "read_only": 1,
+   "remember_last_selected_value": 1,
+   "reqd": 1
   },
   {
    "default": "0",
-   "fieldname": "affects_sales_month",
+   "fieldname": "affects_purchases_book",
    "fieldtype": "Check",
-   "label": "Affects Sales Month"
+   "label": "Affects Purchases Book"
+  },
+  {
+   "default": "0",
+   "fieldname": "affects_purchases_month",
+   "fieldtype": "Check",
+   "label": "Affects Purchases Month"
+  },
+  {
+   "fieldname": "supplier",
+   "fieldtype": "Link",
+   "in_list_view": 1,
+   "label": "Supplier",
+   "options": "Supplier",
+   "reqd": 1
   },
   {
    "fieldname": "amended_from",
@@ -149,40 +157,13 @@
    "read_only": 1
   },
   {
-   "collapsible": 1,
-   "fieldname": "cai_info_section",
-   "fieldtype": "Section Break",
-   "label": "Cai Info"
-  },
-  {
-   "fieldname": "cai",
-   "fieldtype": "Data",
-   "label": "Cai",
-   "read_only": 1
-  },
-  {
-   "fieldname": "due_date_cai",
-   "fieldtype": "Date",
-   "label": "Due Date Cai",
-   "read_only": 1
-  },
-  {
-   "fieldname": "authorized_range",
-   "fieldtype": "Data",
-   "label": "Authorized Range",
-   "read_only": 1
-  },
-  {
    "fieldname": "taxes",
    "fieldtype": "Table",
    "label": "Taxes",
    "options": "Multiple Taxes"
   },
   {
-   "fieldname": "section_break_3",
-   "fieldtype": "Section Break"
-  },
-  {
+   "fetch_from": "taxes.base_isv * (taxes.tax_rate/100)",
    "fieldname": "isv_15",
    "fieldtype": "Currency",
    "label": "Isv 15%",
@@ -190,6 +171,7 @@
    "read_only": 1
   },
   {
+   "fetch_from": "taxes.base_isv * (taxes.tax_rate/100)",
    "fieldname": "isv_18",
    "fieldtype": "Currency",
    "label": "Isv 18%",
@@ -197,15 +179,12 @@
    "read_only": 1
   },
   {
-   "fieldname": "amount_total",
-   "fieldtype": "Currency",
-   "label": "Amount Total",
-   "options": "Currency",
-   "read_only": 1
+   "fieldname": "section_break_3",
+   "fieldtype": "Section Break"
   }
  ],
  "is_submittable": 1,
- "modified": "2021-05-04 23:23:14.796802",
+ "modified": "2021-05-07 11:26:32.489311",
  "modified_by": "Administrator",
  "module": "Accounts",
  "name": "Credit Note CXP",
